+++ jug/tests/jugfiles/simple.py
@@ -13,7 +13,8 @@ def plus1(x):
     return x + 1
 
 vals = list(range(8))
-vals = list(map(double, vals))
-vals = list(map(plus1, vals))
+
+vals = [double(v) for v in vals]
+vals = [plus1(v) for v in vals]
 vals = [sum2(v, 2) for v in vals]
 vals = [sum2(v, v) for v in vals]